bot danteslto/requirements.txt
twitchio
requests
python-dotenv